+++ sdk/keyvault/keyvault-keys/tests/CRUD.test.ts
@@ -2,54 +2,26 @@
 // Licensed under the MIT License.
 
 import * as assert from "assert";
-import { getKeyvaultName } from "./utils/utils.common";
 import { KeysClient, CreateEcKeyOptions, UpdateKeyOptions, GetKeyOptions } from "../src";
 import { RestError } from "@azure/core-http";
-import { EnvironmentCredential } from "@azure/identity";
-import {
-  record,
-  setReplaceableVariables,
-  retry,
-  setReplacements,
-  env,
-  uniqueString
-} from "./utils/recorder";
+import { retry, env } from "./utils/recorder";
+import { authenticate } from "./utils/testAuthentication";
 import TestClient from "./utils/testClient";
 import { AbortController } from "@azure/abort-controller";
 
 describe("Keys client - create, read, update and delete operations", () => {
+  const keyPrefix = `recover${env.KEY_NAME || "KeyName"}`;
+  let keySuffix: string;
   let client: KeysClient;
   let testClient: TestClient;
   let recorder: any;
 
-  const keyPrefix = `CRUD${env.KEY_NAME || "KeyName"}`;
-  let keySuffix: string;
-
   before(async function() {
-    // NOTE:
-    // setReplaceableVariables and setReplacements are reused just to put their ussage in the open,
-    // to avoid having them obscured into a generic utility file. Once the recording tool is centralized
-    // we can move these somewhere else!
-    setReplaceableVariables({
-      AZURE_CLIENT_ID: "azure_client_id",
-      AZURE_CLIENT_SECRET: "azure_client_secret",
-      AZURE_TENANT_ID: "azure_tenant_id",
-      KEYVAULT_NAME: "keyvault_name"
-    });
-
-    keySuffix = uniqueString();
-    setReplacements([
-      (recording: any): any => recording.replace(/"access_token":"[^"]*"/g, `"access_token":"access_token"`),
-      (recording: any): any =>
-        keySuffix === "" ? recording : recording.replace(new RegExp(keySuffix, "g"), "")
-    ]);
-
-    recorder = record(this); // eslint-disable-line no-invalid-this
-    const credential = await new EnvironmentCredential();
-    const keyVaultName = getKeyvaultName();
-    const keyVaultUrl = `https://${keyVaultName}.vault.azure.net`;
-    client = new KeysClient(keyVaultUrl, credential);
-    testClient = new TestClient(client);
+    const authentication = await authenticate(this); // eslint-disable-line no-invalid-this
+    keySuffix = authentication.keySuffix;
+    client = authentication.client;
+    testClient = authentication.testClient;
+    recorder = authentication.recorder;
   });
 
   after(async function() {
